+++ lib/core/widgets/lazy_indexed_stack.dart
@@ -1,3 +1,4 @@
+import 'package:counter/core/p0u_startup_diag.dart';
 import 'package:flutter/material.dart';
 /// [IndexedStack] that lazily builds children on first visit and pauses
 /// off-screen tabs (no paint, no tickers) for smoother shell navigation.
@@ -16,6 +17,16 @@ class LazyIndexedStack extends StatefulWidget {
 }
 
 class _LazyIndexedStackState extends State<LazyIndexedStack> {
+  static bool _bootDeferLogged = false;
+  static const List<String> _tabNames = [
+    'Timeline',
+    'Plans',
+    'Calendar',
+    'Lists',
+    'Categories',
+    'Profile',
+  ];
+
   late List<bool> _visited;
 
   @override
@@ -23,6 +34,17 @@ class _LazyIndexedStackState extends State<LazyIndexedStack> {
     super.initState();
     _visited = List<bool>.filled(widget.children.length, false);
     _markVisited(widget.index);
+    if (!_bootDeferLogged) {
+      _bootDeferLogged = true;
+      for (var i = 0; i < _tabNames.length && i < widget.children.length; i++) {
+        if (i != widget.index) {
+          P0uStartupDiag.hiddenTabDeferred(
+            tab: _tabNames[i],
+            reason: 'firstFramePriority',
+          );
+        }
+      }
+    }
   }
 
   @override
@@ -35,7 +57,23 @@ class _LazyIndexedStackState extends State<LazyIndexedStack> {
       }
       _visited = next;
     }
-    _markVisited(widget.index);
+    final idx = widget.index;
+    final wasVisited = idx >= 0 && idx < _visited.length && _visited[idx];
+    _markVisited(idx);
+    if (!wasVisited &&
+        idx >= 0 &&
+        idx < _visited.length &&
+        _visited[idx] &&
+        idx < _tabNames.length) {
+      final sw = Stopwatch()..start();
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        sw.stop();
+        P0uStartupDiag.hiddenTabActivated(
+          tab: _tabNames[idx],
+          ms: sw.elapsedMilliseconds,
+        );
+      });
+    }
   }
 
   void _markVisited(int i) {